+++ composer.json
@@ -87,6 +87,7 @@
             "Pehapkari\\": "src",
             "Pehapkari\\Provision\\": "packages/Provision/src",
             "Pehapkari\\Registration\\": "packages/Registration/src",
+            "Pehapkari\\Github\\": "packages/Github/src",
             "Pehapkari\\Training\\": "packages/Training/src",
             "Pehapkari\\Marketing\\": "packages/Marketing/src",
             "Pehapkari\\BetterEasyAdmin\\": "packages/BetterEasyAdmin/src",
